Text Field movement
Very small addition to make sure all text fields have the capability to move to the next text field after pressing the android keyboard's "enter" key. Fixed the issue of creating new lines on the text fields that didn't have this feature.  (Looking into a requestFocus feature as well for the first EditText)